Move errors and interfaces to separate files
