+++ restapi/service/auth.go
@@ -8,24 +8,15 @@ import (
 	"github.com/jmoiron/sqlx"
 	"github.com/kngnkg/tunetrail/restapi/auth"
 	"github.com/kngnkg/tunetrail/restapi/model"
-	"github.com/kngnkg/tunetrail/restapi/store"
 )
 
-// var (
-// // エラー
-// )
-
 type AuthService struct {
-	DB   store.Beginner
+	DB   *sqlx.DB
 	Repo UserRepository
 	Auth Auth
 }
 
 // type Token struct{}
-
-type Auth interface {
-	SignUp(ctx context.Context, email, password string) (string, error)
-}
 
 // RegisterUserはユーザーを登録する
 func (as *AuthService) RegisterUser(ctx context.Context, data *model.UserRegistrationData) (*model.User, error) {
