+++ restapi/service/health.go
@@ -9,10 +9,6 @@ import (
 	"github.com/kngnkg/tunetrail/restapi/store"
 )
 
-type HealthRepository interface {
-	Ping(ctx context.Context, db store.Queryer) error
-}
-
 type HealthService struct {
 	DB   store.Queryer
 	Repo HealthRepository
