+++ restapi/service/user.go
@@ -10,31 +10,6 @@ import (
 	"github.com/kngnkg/tunetrail/restapi/store"
 )
 
-var (
-	// ユーザー名が既に存在する
-	ErrUserNameAlreadyExists = errors.New("service: user name already exists")
-	// メールアドレスが既に存在する
-	ErrEmailAlreadyExists = errors.New("service: email already exists")
-	// ユーザーが存在しない
-	ErrUserNotFound = errors.New("service: user not found")
-)
-
-type UserRepository interface {
-	// WithTransactionはトランザクションを実行する
-	WithTransaction(ctx context.Context, db store.Beginner, f func(tx *sqlx.Tx) error) error
-	// RegisterUserはユーザーを登録する
-	// RegisterUserInfo
-	RegisterUser(ctx context.Context, db store.Queryer, u *model.User) error
-	// UserExistsByUserNameはユーザー名が既に存在するかどうかを返す
-	UserExistsByUserName(ctx context.Context, db store.Queryer, userName string) (bool, error)
-	// GetUserByUserNameはユーザー名からユーザーを取得する
-	GetUserByUserName(ctx context.Context, db store.Queryer, userName string) (*model.User, error)
-	// UpdateUserはユーザーを更新する
-	UpdateUser(ctx context.Context, db store.Queryer, u *model.User) error
-	// DeleteUserByUserNameはユーザー名からユーザーを削除する
-	DeleteUserByUserName(ctx context.Context, db store.Queryer, userName string) error
-}
-
 type UserService struct {
 	DB   store.Beginner
 	Repo UserRepository
